getting rid of divide by 0 error messages for log fold change (#46)
Got rid of dividing by 0 errors and invalid error (when 0/0). Currently, if control is 0, the lfc will be inf (now without an error message). Keep in mind that if treatment is 0, lfc will be -inf, and if they are both zero, the value becomes NaN (now without an error message) . For visualizing lfc with a histogram, these should be removed.